Move docker files to superset dir
