+++ srcd/contrib/docker/add_gitbase.py
